cleaning block manager
